Changed default vim theme to blue
gVim looked fine with solarized, but solarized in a 16-color terminal looks
kind of funny. I chose blue because it looked best in the terminal, because I
use vim more than gVim. If I really sit down to code, I use Emacs, but if I'm
editing on the fly and hopping file to file, vim is what I use.