Cleaned .emacs
It was annoying me to switch themes and get "do you really want to load this
theme with lisp code" and "add this to white list", but apparently that all can
be bypassed in Emacs 24 by calling (load-theme 'theme t). I did so, and removed
the white list variables from the custom set vars so there's less clutter.