Cleaned .vimrc
Vim was seeming to ignore my rc for ruby files, but it was because ruby.vim
didn't exist. I had to add that to my ~/.vim/ftplugin/ dir. Also changed
fileformats= to fileformat=unix because that's what I'm working in, and I can
always convert it if I absolutely have to.