Fiddled with RVM PATH lines
RVM is throwing a fit over my default ruby not being in the path. I've tried
everything I can think of, but the only solution I have so far is to call
default ruby, so I went ahead and made my zshrc call it silently.

Still trying to get my .irbrc to look nice. Every time I require a gem in it, I
get a LoadError, and it's apparetly because of how RVM handles gemsets. Ugh.