e => 'emacs --no-window' (run in terminal)
One little reason I moved back to vim from Emacs was that calling
emacs from the terminal on my dev computer opened the GUI version
by default, where when doing the same on my Windows machine the
terminal version was opened. If I'm in a terminal, chances are I
want to be in a terminal, Emacs.

And, like magic, Emacs' solarized theme looks for the most part
like vim's, only differing because its Ruby mode doesn't highlight
iterators or numbers. That's sad, but I can live with that.