Moved Caps Lock and Tap settings to autostart
LXDE respects most, but not all of lxderc.sh. It'll start Dropbox
like I ask in lxderc.sh, and gnome-terminal full screen like I ask
in lxderc.sh, but when I start working, Caps is still Caps and I
still have Tap-To-Click on. My guess is that the defaults are read
after running all the scripts referenced in autostart. If that is
the case, this should fix it, the commands being at the end of the
startup script.

If not, well, I'll make another script I can run at the start myself,
since running the old lxderc.sh myself will set the caps and tap
settings like I wanted.