fiddled with themes
Terminal emacs doesn't properly load a chosen color-theme unless
it is done after .emacs has been evaluated. Hooray for fixing a
race condition with a 'wait 1'.

It's probably something to do with my .emacs, but it does not
happen in gui-emacs, and thus I didn't issue a wait command for
gui-emacs.